Bug Fix: Location of Toast Icons
Previously, the script was looking for the toast icons in $pwd.
But if we were to execute the script by its absolute path from another
directory, then $pwd != location-of-the-script, so we would be looking for
the icons in the long location.
To fix this, I am instead looking in $PSScriptRoot, which refers to the
directory in which the SCRIPT is located.